Add eslint and update all files in src/
Setup a npm script to lint with: `npm run lint`

Using airbnb's eslint config as well as some stylistic ones. Updated all
files to pass linting rules. Threw in some TODOs and ignores that will
need to be double checked.

This commit just does the src/ directory, will add the rest in a later
commit, but this is a good start. It will also run the lint task before
running the tests.