Add UMD building as npm tasks
This should use the existing version of webpack and
configuration for babel to build an UMD library version
of React Look.

The React package is externalised. This means that if
the UMD file is used, React must be already present
as either a commonjs2 or amd import or in the global
root var as 'React'.

Lodash on the other hand is not externalised, as the
babel plugin takes care of including only the methods
that are actually in use.